fixed docstrings
